refactor(web): extract getLineChanges as function
